# public/openapi.yaml
openapi: 3.0.0
info:
  title: Iran Memorial API
  version: 1.0.0
  description: |
    Open API for accessing Iran Memorial data — victims, events, statistics, and sources.

    **Authentication:** All endpoints require an API key via `Authorization: Bearer iran_mem_...` header.

    **Rate Limits:** Default 1000 requests/hour per API key.

    **License:** Data is CC BY-SA 4.0. Attribution required.
  contact:
    email: <CONTACT_EMAIL>
  license:
    name: CC BY-SA 4.0
    url: https://creativecommons.org/licenses/by-sa/4.0/
servers:
  - url: <DEPLOYMENT_URL>/api/v1
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /victims/{slug}:
    get:
      summary: Get victim by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Success
        '404':
          description: Victim not found
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: iran_mem_...
      description: API key with format iran_mem_XXXX
...